+++ utils/retry/retry_configuration_test.go
@@ -7,6 +7,7 @@ package retry
 import (
 	"testing"
 
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -23,6 +24,24 @@ func TestBasicRetryConfiguration(t *testing.T) {
 func TestBasicRetryWithRetryAfterConfiguration(t *testing.T) {
 	configTest := DefaultRobustRetryPolicyConfiguration()
 	require.NoError(t, configTest.Validate())
+	assert.False(t, configTest.RetryAfterDisabled)
+}
+
+func TestFixedBackoffRetryConfiguration(t *testing.T) {
+	configTest := DefaultFixedBackoffRetryPolicyConfiguration()
+	require.NoError(t, configTest.Validate())
+	assert.True(t, configTest.Enabled)
+	assert.True(t, configTest.BackOffEnabled)
+	assert.True(t, configTest.LinearBackOffEnabled)
+	assert.Equal(t, configTest.RetryWaitMin, configTest.RetryWaitMax)
+	assert.Zero(t, configTest.RetryMaxJitter)
+}
+
+func TestRobustFixedBackoffRetryConfiguration(t *testing.T) {
+	configTest := DefaultRobustFixedBackoffRetryPolicyConfiguration()
+	require.NoError(t, configTest.Validate())
+	assert.False(t, configTest.RetryAfterDisabled)
+	assert.True(t, configTest.Enabled)
 }
 
 func TestExponentialBackoffRetryConfiguration(t *testing.T) {